# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Messenger API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создать чат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Создать чат",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user-id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные для создания чата",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddChat"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Чат создан"
                    },
                    "400": {
                        "description": "Запрос не правильно составлен",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Возникла внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chat/add/members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавить участника",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Добавить участника",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user-id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Список users_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddMember"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Чат создан"
                    },
                    "400": {
                        "description": "Запрос не правильно составлен",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Возникла внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chat/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список чатов пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Получить список чатов пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user-id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.Chat"
                            }
                        }
                    },
                    "400": {
                        "description": "Запрос не правильно составлен",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Возникла внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chat/members/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список участников чата",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Получить список участников чата",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.UserInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Запрос не правильно составлен",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Возникла внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chat/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить чат по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Получить чат по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.Chat"
                        }
                    },
                    "400": {
                        "description": "Запрос не правильно составлен",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Возникла внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить чат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Удалить чат",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Чат удален"
                    },
                    "400": {
                        "description": "Запрос не правильно составлен",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Возникла внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.Chat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isDirect": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/core.User"
                },
                "ownerId": {
                    "type": "integer"
                }
            }
        },
        "core.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "core.UserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "handler.AddChat": {
            "type": "object",
            "properties": {
                "is_direct": {
                    "type": "boolean"
                },
                "members_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.AddMember": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer"
                },
                "members_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "resp.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Укажите 'Bearer', а затем ваш JWT токен.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}